Trail1/requirements.txt
torch>=1.12.0
torchvision>=0.13.0
timm>=0.6.12
scikit-learn
tqdm
Pillow
numpy
streamlit
pandas
matplotlib
seaborn
